refactor(login): moved from pages to features, added getLoginUrl export (missing from prev commit)
